backend/requirements.txt
python-dotenv==1.0.0
gql[requests]==3.5.0
requests==2.31.0

